+++ src/main/java/com/hh/mirishop/productmanagement/product/dto/ProductCreate.java
@@ -1,4 +1,4 @@
-package com.hh.mirishop.productmanagement.dto;
+package com.hh.mirishop.productmanagement.product.dto;
 
 
 import lombok.Getter;
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class ProductUpdate {
+public class ProductCreate {
 
     private String name;
     private String content;
