+++ src/main/java/com/hh/mirishop/productmanagement/product/dto/ProductResponse.java
@@ -1,6 +1,6 @@
-package com.hh.mirishop.productmanagement.dto;
+package com.hh.mirishop.productmanagement.product.dto;
 
-import com.hh.mirishop.productmanagement.entity.Product;
+import com.hh.mirishop.productmanagement.product.entity.Product;
 import lombok.Getter;
 
 @Getter
