+++ src/main/java/com/hh/mirishop/productmanagement/product/service/ProductQueryService.java
@@ -1,6 +1,6 @@
-package com.hh.mirishop.productmanagement.service;
+package com.hh.mirishop.productmanagement.product.service;
 
-import com.hh.mirishop.productmanagement.dto.ProductResponse;
+import com.hh.mirishop.productmanagement.product.dto.ProductResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
