+++ src/main/java/com/hh/mirishop/productmanagement/product/service/ProductService.java
@@ -1,0 +1,13 @@
+package com.hh.mirishop.productmanagement.product.service;
+
+import com.hh.mirishop.productmanagement.product.dto.ProductCreate;
+import com.hh.mirishop.productmanagement.product.dto.ProductUpdate;
+
+public interface ProductService {
+
+    Long create(ProductCreate productCreate);
+
+    void update(Long productId, ProductUpdate productUpdate);
+
+    void delete(Long productId);
+}
